claude: refactor(kaspa): move bridge module from libs/kaspa to rust/main/dymension-kaspa
Move the bridge crate from libs/kaspa/lib/bridge to
rust/main/chains/dymension-kaspa/src/kas_bridge.

This continues the refactoring effort to separate pure Kaspa code
(libs/kaspa) from Hyperlane integration code (rust/main). The bridge
module contains HL-dependent types for deposit/withdraw/confirmation
payloads, message parsing, and user-facing deposit functionality.

Changes:
- Move bridge module files to dymension-kaspa as kas_bridge
- Update imports from prost to hyperlane_cosmos_rs::prost
- Update dym_kas_bridge:: references to crate::kas_bridge::
- Update secp256k1 imports to kaspa_bip32::secp256k1
- Add workflow-core to workspace dependencies
- Remove bridge from libs/kaspa workspace members
- Update tooling to use dymension_kaspa::kas_bridge::
