+++ dymension/libs/kaspa/Cargo.toml
@@ -4,7 +4,6 @@ resolver = "2" # Enables new feature resolver, good practice for new workspaces
 
 members = [
     "lib/core",
-    "lib/bridge",
     "lib/api",
     "lib/hardcode",
     "lib/kms",
