--- dymension/libs/kaspa/lib/bridge/src/mod.rs
@@ -1,7 +1,0 @@
-pub mod confirmation;
-pub mod deposit;
-pub mod message;
-pub mod payload;
-pub mod user;
-pub mod util;
-pub mod withdraw;
