+++ dymension/libs/kaspa/tooling/src/x/deposit.rs
@@ -1,4 +1,4 @@
-use bridge::user::deposit::deposit_with_payload;
+use dymension_kaspa::kas_bridge::user::deposit::deposit_with_payload;
 use corelib::wallet::get_wallet;
 use eyre::Result;
 use kaspa_addresses::Address;
