+++ rust/main/chains/dymension-kaspa/src/kas_bridge/confirmation.rs
@@ -1,4 +1,4 @@
-use crate::payload::MessageID;
+use crate::kas_bridge::payload::MessageID;
 use bytes::Bytes;
 use eyre::Error as EyreError;
 use hex::ToHex;
