+++ rust/main/chains/dymension-kaspa/src/kas_bridge/payload.rs
@@ -2,7 +2,7 @@ use bytes::Bytes;
 use eyre::Error as EyreError;
 use hyperlane_core::{Encode, HyperlaneMessage, H256};
 use hyperlane_cosmos_rs::dymensionxyz::hyperlane::kaspa::MessageIDs as ProtoMessageIDs;
-use prost::Message;
+use hyperlane_cosmos_rs::prost::Message;
 
 #[derive(Debug, Clone, PartialEq, Eq, Copy)]
 pub struct MessageID(pub H256);
