+++ rust/main/chains/dymension-kaspa/src/kas_relayer/withdraw/messages.rs
@@ -2,14 +2,16 @@ use super::hub_to_kaspa::{
     build_withdrawal_pskt, extract_current_anchor, fetch_input_utxos, get_normal_bucket_feerate,
     get_outputs_from_msgs,
 };
-use crate::kas_relayer::withdraw::sweep::{create_inputs_from_sweeping_bundle, create_sweeping_bundle};
+use crate::kas_bridge::payload::MessageIDs;
+use crate::kas_bridge::withdraw::{filter_pending_withdrawals, WithdrawFXG};
+use crate::kas_relayer::withdraw::sweep::{
+    create_inputs_from_sweeping_bundle, create_sweeping_bundle,
+};
 use dym_kas_core::consts::RELAYER_SIG_OP_COUNT;
 use dym_kas_core::escrow::EscrowPublic;
-use dym_kas_bridge::payload::MessageIDs;
 use dym_kas_core::wallet::EasyKaspaWallet;
-use dym_kas_bridge::withdraw::{filter_pending_withdrawals, WithdrawFXG};
-use eyre::Result;
 use dym_kas_hardcode::tx::{MAX_MASS_MARGIN, SWEEPING_THRESHOLD};
+use eyre::Result;
 use hyperlane_core::HyperlaneMessage;
 use hyperlane_core::U256;
 use hyperlane_cosmos::{native::ModuleQueryClient, CosmosProvider};
