+++ rust/main/chains/dymension-kaspa/src/kas_validator/confirmation.rs
@@ -1,7 +1,7 @@
+use crate::kas_bridge::confirmation::ConfirmationFXG;
+use crate::kas_bridge::payload::{MessageID, MessageIDs};
 use crate::kas_validator::error::ValidationError;
 use dym_kas_api::models::{TxModel, TxOutput};
-use dym_kas_bridge::confirmation::ConfirmationFXG;
-use dym_kas_bridge::payload::{MessageID, MessageIDs};
 use dym_kas_core::api::client::HttpClient;
 use dym_kas_core::finality::is_safe_against_reorg;
 use hyperlane_cosmos_rs::dymensionxyz::dymension::kas::ProgressIndication;
