+++ rust/main/chains/dymension-kaspa/src/lib.rs
@@ -21,11 +21,11 @@ mod endpoints;
 mod util;
 mod withdrawal_utils;
 
+pub mod kas_bridge;
 pub mod kas_relayer;
 pub mod kas_validator;
 
 // Direct reexports of lib stuff:
-pub use dym_kas_bridge as kas_bridge;
 pub use dym_kas_core;
 pub use dymension_kaspa_hl_constants as hl_domains;
