+++ rust/main/chains/dymension-kaspa/src/providers/provider.rs
@@ -1,17 +1,17 @@
 use super::confirmation_queue::PendingConfirmation;
 use super::validators::ValidatorsClient;
 use super::RestProvider;
+use crate::kas_bridge::confirmation::ConfirmationFXG;
+use crate::kas_relayer::withdraw::hub_to_kaspa::combine_bundles_with_fee;
+use crate::kas_relayer::withdraw::messages::on_new_withdrawals;
+use crate::kas_relayer::KaspaBridgeMetrics;
 use crate::util::domain_to_kas_network;
 use crate::withdrawal_utils::{record_withdrawal_batch_metrics, WithdrawalStage};
 use crate::ConnectionConf;
 use crate::RelayerStuff;
 use crate::ValidatorStuff;
-use crate::kas_bridge::confirmation::ConfirmationFXG;
 use dym_kas_core::escrow::EscrowPublic;
 use dym_kas_core::wallet::{EasyKaspaWallet, EasyKaspaWalletArgs};
-use crate::kas_relayer::withdraw::hub_to_kaspa::combine_bundles_with_fee;
-use crate::kas_relayer::withdraw::messages::on_new_withdrawals;
-use crate::kas_relayer::KaspaBridgeMetrics;
 use eyre::Result;
 use hyperlane_core::config::OpSubmissionConfig;
 use hyperlane_core::NativeToken;
